feat: implement BaseClusterParser for shared clustering functionality
- Introduced BaseClusterParser class to handle parsing of MMseqs2 clustering outputs.
- Added methods for splitting representatives into training, validation, and test sets.
- Refactored EasyClusterParser and EasyLinClustParser to inherit from BaseClusterParser, promoting code reuse.
- Updated tests to validate the new functionality and ensure compatibility with existing configurations.
